기능 구현 완료입니다

+++ src/main/java/com/mycompany/project/reservation/query/dto/ReservationDTO.java
@@ -4,17 +4,16 @@ import lombok.*;
 
 import java.time.LocalDateTime;
 
-
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
 public class ReservationDTO {
-  private int reservationId;
-  private LocalDateTime startTime;
-  private LocalDateTime endTime;
-  private String status;
-  private LocalDateTime createdAt;
-  private String rejectionReason;
+    private Long reservationId;
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
+    private String status;
+    private LocalDateTime createdAt;
+    private String rejectionReason;
 }
